Disable support for systems needing 64-bit indices
The 64-bit support isn't implemented everywhere and probably would exceed
memory limits even if it was. BigUINT has been set to 32 bit and a check
is added to cudaMain to confirm that this is large enough for the system.
If it's not, the program will exit with a failure status.